feat(penilaiancontroller): menambahkan fitur metod edit and update

+++ database/migrations/2026_05_01_072707_create_detail_penilaians_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('penilaian_id')->constrained()->cascadeOnDelete();
             $table->foreignId('karyawan_id')->constrained()->cascadeOnDelete();
             $table->foreignId('kriteria_id')->constrained()->cascadeOnDelete();
-            $table->float('nilai');
+            $table->integer('nilai');
             $table->timestamps();
         });
     }
